fixing bogs

+++ tanks/Transition.h
@@ -4,7 +4,7 @@ class Transition : public Scene {
 private:
 	Scene* scene;
 	double y = 1000;
-	double zoom = 0.5;
+	float zoom = 0.5;
 public:
 	Transition(Scene* scene);
